+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - Properly handle invalid value for `NSUnderlyingErrorKey` (#4144)
+- Session replay in buffer mode not working ()
 
 ## 8.30.1